fix: clippy::module_inception
