+++ abci/src/server/unix.rs
@@ -2,7 +2,7 @@
 
 use std::{fs::remove_file, path::Path};
 
-use crate::{server::server::handle_client, Application, Error};
+use super::{handle_client, Application, Error};
 use std::os::unix::net::UnixListener;
 use tracing::info;
